Seeded inital data to ease testing. - 3 roles - 4 users (1 admin, 1 publisher, 2 users) - 1 book published by publisher - 2 reviews left by 2 users

+++ Bookie.Domain/Entities/Book.cs
@@ -18,7 +18,7 @@ namespace Bookie.Domain.Entities
         public string Language { get; set; }
         public string CoverPhotoUrl { get; set; }
         public DateTime CreatedAt { get; set; }
-        public decimal AverageRating { get; set; }
+        public double AverageRating { get; set; }
         public string GetBookUrl { get; set; }
 
         public Guid? GenreId { get; set; }
